requirements.txt
metaflow==2.2.13
numpy==1.19.5
tensorflow==2.5.0
python-dotenv==0.13.0
scikit-learn==0.24.2
wandb==0.10.30